Added the roles in all the routers endpoints and services
Did some code formatting.
Added academic_year in permission checks.
Added all the correct get permissions for each services.
Added some missing documentation.

+++ app/services/RoleService.py
@@ -12,17 +12,19 @@ from app.models.tortoise.permission import PermissionInDB
 from app.models.tortoise.role import RoleInDB
 from app.services.PermissionService import check_permissions
 from app.utils.enums.http_errors import CommonErrorMessages
-from app.utils.enums.permission_enums import AvailableServices, AvailableOperations
+from app.utils.enums.permission_enums import AvailableRoles, AvailableServices, AvailableOperations
 
 
-async def get_all_roles(current_account: AuthenticatedAccount) -> list[PydanticRoleResponseModel]:
+async def get_all_roles(academic_year: int,
+                        current_account: AuthenticatedAccount) -> list[PydanticRoleResponseModel]:
     """
     This method retrieves all roles.
     """
 
     await check_permissions(AvailableServices.ACCOUNT_SERVICE,
-                            AvailableOperations.GET,
-                            current_account)
+                            AvailableOperations.GET_MULTIPLE,
+                            current_account,
+                            academic_year)
 
     roles: list[RoleInDB] = await RoleInDB.all()
     roles_list : list[PydanticRoleResponseModel] = []
@@ -37,19 +39,23 @@ async def get_all_roles(current_account: AuthenticatedAccount) -> list[PydanticR
     return roles_list
 
 
-async def get_role_by_id(name: str, current_account: AuthenticatedAccount) -> PydanticRoleResponseModel:
+async def get_role_by_id(academic_year: int,
+                         name: str,
+                         current_account: AuthenticatedAccount) -> PydanticRoleResponseModel:
     """
-        This method retrieves a role with given name.
+    This method retrieves a role with given name.
     """
 
     await check_permissions(AvailableServices.ACCOUNT_SERVICE,
-                            AvailableOperations.GET,
-                            current_account)
+                            AvailableOperations.GET_SINGLE,
+                            current_account,
+                            academic_year)
 
     role: RoleInDB | None = await RoleInDB.get_or_none(name=name)
 
     if role is None:
-        raise HTTPException(status_code=404, detail=CommonErrorMessages.ROLE_NOT_FOUND.value)
+        raise HTTPException(status_code=404,
+                            detail=CommonErrorMessages.ROLE_NOT_FOUND.value)
 
     return PydanticRoleResponseModel(
         name=role.name,
@@ -57,14 +63,17 @@ async def get_role_by_id(name: str, current_account: AuthenticatedAccount) -> Py
     )
 
 
-async def add_role(role: PydanticCreateRoleModel, current_account: AuthenticatedAccount) -> None:
+async def add_role(academic_year: int,
+                   role: PydanticCreateRoleModel,
+                   current_account: AuthenticatedAccount) -> None:
     """
-        This method delete a role with given name.
+    This method delete a role with given name.
     """
 
     await check_permissions(AvailableServices.ACCOUNT_SERVICE,
                             AvailableOperations.CREATE,
-                            current_account)
+                            current_account,
+                            academic_year)
 
     if await RoleInDB.filter(name=role.name).exists():
         raise HTTPException(status_code=409, detail=CommonErrorMessages.ROLE_ALREADY_EXIST.value)
@@ -77,39 +86,56 @@ async def add_role(role: PydanticCreateRoleModel, current_account: Authenticated
         for permission in role.permissions:
             perm: PermissionInDB | None = await PermissionInDB.get_or_none(id=permission)
             if not perm:
-                raise HTTPException(status_code=404, detail=CommonErrorMessages.PERMISSION_NOT_FOUND.value)
+                raise HTTPException(status_code=404,
+                                    detail=CommonErrorMessages.PERMISSION_NOT_FOUND.value)
             await role_to_create.permissions.add(perm)
 
 
-async def modify_role(role_name : str, body: PydanticUpdateRoleModel , current_account: AuthenticatedAccount) -> None:
+async def modify_role(academic_year: int,
+                      role_name : str,
+                      body: PydanticUpdateRoleModel,
+                      current_account: AuthenticatedAccount) -> None:
     """
     This method modifies a role with given name.
     """
 
     await check_permissions(AvailableServices.ACCOUNT_SERVICE,
                             AvailableOperations.UPDATE,
-                            current_account)
+                            current_account,
+                            academic_year)
 
     role: RoleInDB | None = await RoleInDB.get_or_none(name=role_name)
 
     if role is None:
         raise HTTPException(status_code=404, detail=CommonErrorMessages.ROLE_NOT_FOUND.value)
 
+    if role.name == AvailableRoles.ADMIN.value.role_name:
+        raise HTTPException(status_code=403,
+                            detail=CommonErrorMessages.CANNOT_UPDATE_ADMIN.value)
+
     return None
 
 
-async def delete_role(role_name: str, current_account: AuthenticatedAccount) -> None:
+async def delete_role(academic_year: int,
+                      role_name: str,
+                      current_account: AuthenticatedAccount) -> None:
     """
     This method delete a role with given name.
     """
 
     await check_permissions(AvailableServices.ACCOUNT_SERVICE,
                             AvailableOperations.DELETE,
-                            current_account)
+                            current_account,
+                            academic_year)
 
     role: RoleInDB | None = await RoleInDB.get_or_none(name=role_name)
 
     if role is None:
-        raise HTTPException(status_code=404, detail=CommonErrorMessages.ROLE_NOT_FOUND.value)
+        raise HTTPException(status_code=404,
+                            detail=CommonErrorMessages.ROLE_NOT_FOUND.value)
+
+    if role.name == AvailableRoles.ADMIN.value.role_name:
+        raise HTTPException(status_code=403,
+                            detail=CommonErrorMessages.CANNOT_DELETE_ADMIN.value)
 
     await role.delete()
